+++ tensorlayer/networks/custom_model.py
@@ -35,22 +35,41 @@ class CustomModel(BaseNetwork, ABC):
         else:
             input_layer, output_layer = model
 
-        if not isinstance(
-            input_layer,
-            (
-                tl.layers.InputLayer, tl.layers.OneHotInputLayer, tl.layers.Word2vecEmbeddingInputlayer,
-                tl.layers.EmbeddingInputlayer, tl.layers.AverageEmbeddingInputlayer
-            )
-        ):
+        if isinstance(input_layer, (list, tuple)):
+            for layer in input_layer:
+                if layer.__class__.__name__ not in tl.layers.inputs.__all__:
+                    raise RuntimeError(
+                        "The returned input_layer (type: %s) contains a layer (type: %s) which is not "
+                        "a known input layer: %s"
+                        % (type(input_layer), type(layer), tl.layers.inputs.__all__)
+                    )
+
+        elif isinstance(input_layer, tl.layers.Layer):
+            if input_layer.__class__.__name__ not in tl.layers.inputs.__all__:
+                raise RuntimeError(
+                    "The returned input_layer (type: %s) is not an instance of a known input layer: %s" %
+                    (type(input_layer), tl.layers.inputs.__all__)
+                )
+        else:
             raise RuntimeError(
-                "The returned input layer (type: %s) is not an instance of a known input layer: %s" %
-                (type(input_layer), tl.layers.inputs.__all__)
+                "The returned input_layer (type: %s) is not an instance of Layer type or Tuple/List of Layers" %
+                (type(input_layer))
             )
 
-        if not isinstance(output_layer, tl.layers.Layer):
+        if isinstance(output_layer, (list, tuple)):
+            for layer in output_layer:
+                if not isinstance(layer, tl.layers.Layer):
+                    raise RuntimeError(
+                        "The returned output_layer (type: %s) contains a layer (type: %s) which is not "
+                        "an instance of type `tl.layers.Layer`" %
+                        (type(output_layer), type(layer))
+                    )
+
+        elif not isinstance(output_layer, tl.layers.Layer):
             raise RuntimeError(
-                "The returned output layer (type: %s) is not an instance of a `tensorlayer.layers.Layer`" %
-                type(output_layer)
+                "The returned output_layer (type: %s) is not an instance of  type "
+                "`tl.layers.Layer` or Tuple/List of `tl.layers.Layer`" %
+                (type(input_layer))
             )
 
         return input_layer, output_layer